+++ job-search-react/src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
-import JobApp from "./components/JobApp";
+import Header from "./components/Header/Header";
+import Footer from "./components/Footer/Footer";
+import JobApp from "./components/JobApp/JobApp";
 
 function App() {
   return (
