+++ job-search-react/src/components/Footer/Footer.js
