+++ job-search-react/src/components/JobApp/Form/SearchButton/SearchButton.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default function SearchButton() {
+	return (
+		<button
+			type="submit"
+			class="btn btn-primary mb-2 submit-button search">
+			<i class="glyphicon glyphicon-search">Search</i>
+		</button>
+	);
+}
